Add hook useMenu
